% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_trendline.R
\name{geom_trendline}
\alias{geom_trendline}
\title{Add Trendline to 'ggplot'}
\usage{
geom_trendline(mapping = NULL, data = NULL, model = "line2P", ...)
}
\arguments{
\item{model}{select which model to fit. Default is "line2P". The "model" should be one of c("line2P", "line3P", "log2P", "exp3P", "power3P"), their formulas are as follows:\cr "line2P": y=a*x+b \cr "line3P": y=a*x^2+b*x+c \cr "log2P": y=a*ln(x)+b \cr "exp3P": y=a*exp(b*x)+c \cr "power3P": y=a*x^b+c}

\item{...}{additional parameters to \code{\link[ggplot2]{geom_line}},such as color, linetype, size, alpha.}

\item{x, y}{the x and y arguments provide the x and y coordinates for the plot. Any reasonable way of defining the coordinates is acceptable.}
}
\description{
An extension of 'ggplot2' for add trendline to ggplot,
by using different models built in the 'geom_trendline()' function. The function includes the following models in the latest version:
"line2P" (formula as: y=a*x+b), "line3P" (y=a*x^2+b*x+c), "log2P" (y=a*ln(x)+b), "exp3P" (y=a*exp(b*x)+c), and "power3P" (y=a*x^b+c).
}
\details{
The values of each parameter of regression model can be found by typing \code{\link[basicTrendline]{trendline}} fuction in this pacakge.\cr\cr The linear models (line2P, line3P, log2P) in this package are estimated by \code{\link[stats]{lm}} function, while the nonlinear models (exp3P, power3P) are estimated by \code{\link[stats]{nls}} function (i.e., least-squares method).
}
\examples{
library(ggplot2)
library(basicTrendline2)
x <- 1:5
y <- c(2,14,18,19,20)
xy <- data.frame(x, y)

gg <- ggplot(aes(x,y), data = xy) + geom_point()
gg + geom_trendline(model = "line2P")
gg + geom_trendline(model = "log2P", color = "blue")
gg + geom_trendline(model = "exp3P", color="red", linetype="dashed", size = 10, alpha = 0.3)

}
\seealso{
\code{\link{trendline2}}, \code{\link{SSexp3P}}, \code{\link{SSpower3P}}, \code{\link[stats]{nls}}, \code{\link[stats]{selfStart}}
}
\author{
Weiping Mei, Guangchuang Yu
}
